int[]의 int에 float[]의 float이 들어갈 경우 error

+++ test.c
@@ -1,12 +1,12 @@
-float[] change(int[] a) {
-    float b[5];
+int[] change(float[] a) {
+    int b[5];
     b[0] = a[0];
     b[1] = a[1];
     return b;
 }
 
 void main () {
-    int arr[5];
+    float arr[5];
     arr[0] = 600;
     arr[1] = 100;
     _print(change(arr));
